Move structured data to blogpost component
